refactor: 구조 및 컬럼명 변경 반영 - UserBalanceHistory 컬럼명 변경 반영 (type -> status) - infra - repository 구조 변경 반영
